Some data upload refactor
