+++ src/FamilyHubs.ServiceDirectoryAdminUi.Ui/FamilyHubs.ServiceDirectoryAdminUi.Ui.csproj
@@ -11,7 +11,7 @@
     <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.0" />
     <PackageReference Include="BCrypt.Net-Next" Version="4.0.3" />
     <PackageReference Include="DotNetCore.NPOI" Version="1.2.3" />
-    <PackageReference Include="FamilyHubs.ServiceDirectory.Shared" Version="1.1.22" />
+    <PackageReference Include="FamilyHubs.ServiceDirectory.Shared" Version="1.1.25" />
     <PackageReference Include="FluentValidation.AspNetCore" Version="11.2.2" />
     <PackageReference Include="Microsoft.ApplicationInsights" Version="2.21.0" />
     <PackageReference Include="Microsoft.ApplicationInsights.AspNetCore" Version="2.21.0" />
