+++ src/FamilyHubs.ServiceDirectoryAdminUi.Ui/Services/Dataupload/DatauploadService.cs
@@ -1,10 +1,10 @@
 ﻿using FamilyHubs.ServiceDirectory.Shared.Builders;
+using FamilyHubs.ServiceDirectory.Shared.Dto;
 using FamilyHubs.ServiceDirectoryAdminUi.Ui.Models;
 using FamilyHubs.ServiceDirectoryAdminUi.Ui.Pages.OrganisationAdmin;
 using FamilyHubs.ServiceDirectoryAdminUi.Ui.Services.Api;
+using FamilyHubs.ServiceDirectoryAdminUi.Ui.Services.DataUpload.Helpers;
 using System.Data;
-using FamilyHubs.ServiceDirectory.Shared.Dto;
-using FamilyHubs.ServiceDirectory.Shared.Enums;
 
 namespace FamilyHubs.ServiceDirectoryAdminUi.Ui.Services.DataUpload;
 
@@ -16,8 +16,6 @@ public interface IDataUploadService
 
 public class DataUploadService : IDataUploadService
 {
-
-
     private readonly IOrganisationAdminClientService _OrganisationAdminClientService;
     private readonly IPostcodeLocationClientService _postcodeLocationClientService;
 
@@ -62,36 +60,12 @@ public class DataUploadService : IDataUploadService
 
             OrganisationTypeDto organisationTypeDto;
             string? organisationName;
-            switch (dtRow["Organisation Type"].ToString()?.ToLower())
+
+            if(!OrganisationHelper.TryResolveOrganisationType(dtRow, out organisationTypeDto, out organisationName))
             {
-                case "local authority":
-                    organisationTypeDto = new OrganisationTypeDto("1", "LA", "Local Authority");
-                    organisationName = dtRow["Local authority"].ToString();
-                    break;
-                case "voluntary and community sector":
-                    organisationTypeDto = new OrganisationTypeDto("2", "VCFS", "Voluntary, Charitable, Faith Sector");
-                    organisationName = dtRow["Name of organisation"].ToString();
-                    break;
-                case "family hub":
-                    organisationTypeDto = new OrganisationTypeDto("3", "FamilyHub", "Family Hub");
-                    organisationName = dtRow["Local authority"].ToString();
-                    break;
-                default:
-                    organisationTypeDto = new OrganisationTypeDto("4", "Company", "Public / Private Company eg: Child Care Centre");
-                    organisationName = dtRow["Name of organisation"].ToString();
-                    break;
-
+                _errors.Add($"Name of organisation missing row:{rowNumber}");
+                continue;
             }
-
-            if (organisationTypeDto.Name != "LA" && organisationTypeDto.Name != "FamilyHub")
-            {
-                if (string.IsNullOrWhiteSpace(organisationName))
-                {
-                    _errors.Add($"Name of organisation missing row:{rowNumber}");
-                    continue;
-                }
-            }
-
 
             var newOrganisation = false;
             OrganisationWithServicesDto? OrganisationDto;
@@ -234,7 +208,7 @@ public class DataUploadService : IDataUploadService
 
         var builder = new ServicesDtoBuilder();
         var result = builder.WithMainProperties(serviceId,
-                                   GetServiceType(organisationTypeDto),
+                                   ServiceHelper.GetServiceType(organisationTypeDto),
                                    organisationId,
                                    dtRow["Name of service"].ToString() ?? string.Empty,
                                    description,
@@ -246,52 +220,17 @@ public class DataUploadService : IDataUploadService
                                    "active",
                                    string.Empty,
                                    false)
-                        .WithServiceDelivery(GetDeliveryTypes(dtRow["Delivery method"].ToString() ?? string.Empty, service))
+                        .WithServiceDelivery(ServiceHelper.GetDeliveryTypes(dtRow["Delivery method"].ToString() ?? string.Empty, service))
                         .WithServiceAtLocations(locations)
                         .WithLinkContact(ContactHelper.GetLinkContacts(serviceId, LinkContactTypes.SERVICE, dtRow, service?.LinkContacts, _contacts))
-                        .WithCostOption(GetCosts(dtRow, service))
-                        .WithLanguages(GetLanguages(dtRow, service))
+                        .WithCostOption(ServiceHelper.GetCosts(dtRow, service))
+                        .WithLanguages(ServiceHelper.GetLanguages(dtRow, service))
                         .WithServiceTaxonomies(GetTaxonomies(dtRow))
-                        .WithEligibility(GetEligibilities(dtRow, service))
+                        .WithEligibility(ServiceHelper.GetEligibilities(dtRow, service))
                         .Build();
 
         return result;
 
-    }
-
-    private List<EligibilityDto> GetEligibilities(DataRow dtRow, ServiceDto? service)
-    {
-        var eligibilityId = Guid.NewGuid().ToString();
-        var list = (service != null && service.Eligibilities != null) ? service.Eligibilities.ToList() : new();
-
-        if (!int.TryParse(dtRow["Age from"].ToString(), out var minimumAge))
-        {
-            minimumAge = 0;
-        }
-
-        if (!int.TryParse(dtRow["Age to"].ToString(), out var maximumAge))
-        {
-            maximumAge = 127;
-        }
-
-        var eligibility = "Child";
-        if (minimumAge >= 18)
-        {
-            eligibility = "Adult";
-        }
-
-        if (service != null && service.Eligibilities != null)
-        {
-            var eligibleItem = service.Eligibilities?.Count == 1 ? service.Eligibilities?.First() : service.Eligibilities?.FirstOrDefault(x => x.MinimumAge == minimumAge && x.MaximumAge == maximumAge);
-            if (eligibleItem != null)
-            {
-                eligibilityId = eligibleItem.Id;
-            }
-        }
-
-        list.Add(new EligibilityDto(eligibilityId, eligibility, maximumAge, minimumAge));
-
-        return list;
     }
 
     private List<ServiceTaxonomyDto> GetTaxonomies(DataRow dtRow)
@@ -311,125 +250,6 @@ public class DataUploadService : IDataUploadService
 
             }
         }
-        return list;
-    }
-
-    private List<LanguageDto> GetLanguages(DataRow dtRow, ServiceDto? service)
-    {
-        var list = (service != null && service.Languages != null) ? service.Languages.ToList() : new List<LanguageDto>();
-        var languages = dtRow["Language"].ToString();
-        if (!string.IsNullOrEmpty(languages))
-        {
-            var parts = languages.Split('|');
-            foreach (var part in parts)
-            {
-                var languageId = Guid.NewGuid().ToString();
-                if (service != null && service.Languages != null)
-                {
-                    var originalLanguage = service.Languages.FirstOrDefault(x => x.Name == part);
-                    if (originalLanguage != null)
-                    {
-                        languageId = originalLanguage.Id;
-                    }
-                }
-
-                list.Add(new LanguageDto(languageId, part.Trim()));
-            }
-        }
-
-        return list;
-    }
-
-    private List<CostOptionDto> GetCosts(DataRow dtRow, ServiceDto? service)
-    {
-        var list = service?.CostOptions?.Count > 1 ? service.CostOptions.ToList() : new();
-
-        if (string.IsNullOrEmpty(dtRow["Cost (£ in pounds)"].ToString()) &&
-            string.IsNullOrEmpty(dtRow["Cost per"].ToString()) &&
-            string.IsNullOrEmpty(dtRow["Cost Description"].ToString()))
-        {
-            return list;
-        }
-
-        if (!decimal.TryParse(dtRow["Cost (£ in pounds)"].ToString(), out var amount))
-        {
-            amount = 0.0M;
-        }
-
-        var costId = Guid.NewGuid().ToString();
-        if (service != null && service.CostOptions != null)
-        {
-            var costOption = (amount != 0.0M && string.IsNullOrEmpty(dtRow["Cost per"].ToString())) ? service.CostOptions.FirstOrDefault(t => t.Amount == amount && t.AmountDescription == dtRow["Cost per"].ToString()) : service.CostOptions.FirstOrDefault(t => (t.Option == dtRow["Cost Description"].ToString()));
-            if (costOption != null)
-            {
-                costId = costOption.Id;
-            }
-        }
-
-        list.Add(new CostOptionDto(
-                            costId,
-                            dtRow["Cost per"].ToString() ?? string.Empty,
-                            amount,
-                            null,
-                            dtRow["Cost Description"].ToString(),
-                            null,
-                            null
-                            ));
-
-        return list;
-    }
-    private ServiceTypeDto GetServiceType(OrganisationTypeDto organisationTypeDto)
-    {
-        switch (organisationTypeDto.Name)
-        {
-            case "LA":
-            case "FamilyHub":
-                return new ServiceTypeDto("2", "Family Experience", "");
-
-
-            default:
-                return new ServiceTypeDto("1", "Information Sharing", "");
-
-        }
-    }
-
-    private string GetServiceDeliveryId(ServiceDto? service, ServiceDeliveryType serviceDelivery)
-    {
-        var id = Guid.NewGuid().ToString();
-        if (service != null && service.ServiceDeliveries != null)
-        {
-            var serviceDeliveryItem = service.ServiceDeliveries.FirstOrDefault(x => x.Name == serviceDelivery);
-            if (serviceDeliveryItem != null)
-            {
-                id = serviceDeliveryItem.Id;
-            }
-        }
-
-        return id;
-
-    }
-
-    private List<ServiceDeliveryDto> GetDeliveryTypes(string rowDeliveryTypes, ServiceDto? service)
-    {
-        List<ServiceDeliveryDto> list = new();
-        var parts = rowDeliveryTypes.Split('|');
-        foreach (var part in parts)
-        {
-
-            if (string.Compare(part, DeliverMethods.IN_PERSON, StringComparison.OrdinalIgnoreCase) == 0)
-            {
-                list.Add(new ServiceDeliveryDto(GetServiceDeliveryId(service, ServiceDeliveryType.InPerson), ServiceDeliveryType.InPerson));
-            }
-            else if (string.Compare(part, DeliverMethods.ONLINE, StringComparison.OrdinalIgnoreCase) == 0)
-            {
-                list.Add(new ServiceDeliveryDto(GetServiceDeliveryId(service, ServiceDeliveryType.Online), ServiceDeliveryType.Online));
-            }
-            else if (string.Compare(part, DeliverMethods.TELEPHONE, StringComparison.OrdinalIgnoreCase) == 0)
-            {
-                list.Add(new ServiceDeliveryDto(GetServiceDeliveryId(service, ServiceDeliveryType.Telephone), ServiceDeliveryType.Telephone));
-            }
-        }
-
         return list;
     }
